Start on TescoBank Support.
This has the ability to log in. Yay.
At the moment, this does require the php v8js module, as we need to run some Javascript to generate OTPs as part of this login process.
At some point I may look at converting the OTP code from JS to PHP, but for now this seems easier as the JS comes from the HSBC login process and is what a browser would do.

The usual getAccounts() and updateTransactions() functionality to follow.