Add support for accessing HSBC using the api that the mobile client uses.
This allows for unattended updating of HSBC accounts like the rest.

Unfortunately, there are some minor differences with the data that comes
from the mobile API compared to the web:

 - It doesn't provide full account numbers
 - It doesn't provide proper transaction types (just Debit/Credit)
 - Some transactions appear with different wording

If you don't care about these, go ahead and just use HSBCMobile class,
otherwise you will want to use a combination of "HSBCMerge" and "HSBC".

"HSBCMerge" will use the mobile API, but will only update accounts that
are already in the database from "HSBC".

To counter the non-proper transaction types, "importer" now performs an
insert_update with transactions, and "HSBCMerge" reports it's typecode's
as "Generic" where as "HSBC" does not.

"Generic" type codes are removed prior to UPDATE on the database, so when
new transactions are added with "HSBCMerge" they will have the generic
typecode, but if a full update with "HSBC" is performed then these will be
corrected to their real type codes.

To accomodate this, transaction hashcodes no longer include the proper
type codes across the board, they now just use "IN" or "OUT". The
"Database" class will automatically take care of updating the database
when it encounters old-hashed transactions.

As for the different wording, this mostly relates to interest and extra
spaces, which HSBCMobile takes care of, if I find more I'll update it.

Minor changes to debug.php also occurred.

Pressing enter when tagging a transaction now saves the tag.

.gitignore now also ignores "bank_full*.sql" files, to prevent me
accidentally commiting my entire database of transactions...