Add support for new November2016 HSBC UK online banking rewrite.
Hopefully this hasn't broken support for people not yet using the new online
banking, but I can't test this!

The new UI is actually a lot nicer than the old one both for a real person and
for the purposes of this app, as all the data is retrieved from the backend using
a json API. Same as before, CCs randomly do things a bit differently in some
places. I've not figured out how the webapp knows when to do thigns differently
and when not to, unless things break horrible I don't plan to, the web app is
pretty complex javascript (it seems that they use the same app for all the
different countries now and it adapts itself as needed).

Was a bit more effort to get working, but seems to work!

No support for historical data at the moment.

From what I can tell, that redirects to the old-online banking, so should be
easy enough to add support for.