Rework Monzo settlement handling for greater accuracy.
Previously we were only considering settlements after we had considered
everything else, which lead to incorrect handling when there was a long
chain of transactions that needed settlement (such as when using the card
abroad rather than just for one-off alternative-currency transactions)

Now we insert the transactions from the API into a SortedLinkedList and
traverse through this list for the transaction objects.

As we go, if we find a transaction that needs a settlement transaction, we
insert the settlement transaction ahead of us in the list and then parse
it properly when we get to it.

This lets us properly deal with transactions that need settling that occur
after a previous transaction settled.